# server/openapi.spec.yaml
openapi: 3.0.2
info:
  title: DR.CODE Project
  version: 0.0.1
  description: This is the API documentation for DR.CODE Kecskemét's management application
servers:
  - url: http://127.0.0.1:3000
    description: localhost
paths:
...
  /lesson/get:
    get:
      summary: Get lessons by id
      description: Get one or more lesson(s) with the ability to fetch record links.
      tags:
        - lesson
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful and the requested lessons were retrieved.
          content:
            application/json:
              schema:
                type: object
                title: GetLessonsByIdOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lesson"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters:
        - $ref: "#/components/parameters/ids"
      operationId: getLessonsById
  /lesson/between_dates:
    get:
      summary: Get lessons between two dates
      description: Get the lessons starting between two dates.
      tags:
        - lesson
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful and the requested lessons were retrieved.
          content:
            application/json:
              schema:
                type: object
                title: GetLessonsByIdOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lesson"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters:
        - name: start
          in: query
          description: The start date of the range
          required: true
          schema:
            type: string
            format: date-time
            example: 2025-04-01T00:00:00Z
        - name: end
          in: query
          description: The end date of the range
          required: true
          schema:
            type: string
            format: date-time
            example: 2025-04-10T23:59:59Z
      operationId: getLessonsBetweenDates
...
  /lesson/update:
    post:
      summary: Update a lesson
      description: This endpoint can be used to update a lesson as an administrator.
      tags:
        - lesson
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful, and the lesson was updated.
          content:
            application/json:
              schema:
                type: object
                title: UpdateALessonOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - lesson
                    properties:
                      lesson:
                        $ref: "#/components/schemas/Lesson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: updateALesson
      requestBody:
        description: The request body for updating a lesson.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  $ref: "#/components/schemas/ID"
                name:
                  type: string
                  example: Scratch lesson 1
                group:
                  $ref: "#/components/schemas/Group"
                location:
                  $ref: "#/components/schemas/Location"
                teachers:
                  type: array
                  items:
                    $ref: "#/components/schemas/ID"
                start:
                  $ref: "#/components/schemas/Datetime"
                end:
                  $ref: "#/components/schemas/Datetime"
  /lesson/create:
    post:
      summary: Create a new lesson
      description: This endpoint is used to create a new lesson. Only available to admins.
      tags:
        - lesson
      security:
        - BearerJwt: []
      responses:
        "201":
          description: The request was successful, and the lesson was created.
          content:
            application/json:
              schema:
                type: object
                title: CreateALessonCreated
                required:
                  - code
                  - data
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - lesson
                    properties:
                      lesson:
                        $ref: "#/components/schemas/Lesson"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: createALesson
      requestBody:
        description: The request body for creating a new lesson.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - group
                - location
                - teachers
                - start
                - end
              properties:
                name:
                  type: string
                  example: Scratch lesson 1
                group:
                  $ref: "#/components/schemas/Group"
                location:
                  $ref: "#/components/schemas/Location"
                teachers:
                  type: array
                  items:
                    $ref: "#/components/schemas/ID"
                start:
                  $ref: "#/components/schemas/Datetime"
                end:
                  $ref: "#/components/schemas/Datetime"
  # # # # # # # # #
  # INVITE ROUTES #
  # # # # # # # # #
...
  /group/get:
    get:
      summary: Get groups by id
      description: Get one or more group(s) with the ability to fetch record links.
      tags:
        - group
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful and the requested groups were retrieved.
          content:
            application/json:
              schema:
                type: object
                title: GetGroupsByIdOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters:
        - $ref: "#/components/parameters/ids"
      operationId: getGroupsById
...
  /group/update:
    post:
      summary: Update a group
      description: This endpoint can be used to update a group as an administrator.
      tags:
        - group
      security:
        - BearerJwt: []
      responses:
        "200":
          description: The request was successful, and the group was updated.
          content:
            application/json:
              schema:
                type: object
                title: UpdateAGroupOk
                required:
                  - data
                  - code
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - group
                    properties:
                      group:
                        $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: updateAGroup
      requestBody:
        description: The request body for updating a group.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
              properties:
                id:
                  $ref: "#/components/schemas/ID"
                name:
                  type: string
                  example: Kodály 16:00 Scratch
                location:
                  $ref: "#/components/schemas/Location"
                archived:
                  type: boolean
                  description: Whether the group is archived or not
                  example: false
                teachers:
                  type: array
                  items:
                    $ref: "#/components/schemas/ID"
  /group/create:
    post:
      summary: Create a new group
      description: This endpoint is used to create a new group. Only available to admins.
      tags:
        - group
      security:
        - BearerJwt: []
      responses:
        "201":
          description: The request was successful, and the group was created.
          content:
            application/json:
              schema:
                type: object
                title: CreateAGroupCreated
                required:
                  - code
                  - data
                properties:
                  code:
                    type: string
                    example: success
                  data:
                    type: object
                    required:
                      - group
                    properties:
                      group:
                        $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
      parameters: []
      operationId: createAGroup
      requestBody:
        description: The request body for creating a new group.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - location
                - teachers
              properties:
                name:
                  type: string
                  example: Kodály 16:00 Scratch
                location:
                  $ref: "#/components/schemas/Location"
                teachers:
                  type: array
                  items:
                    $ref: "#/components/schemas/ID"
                lessons:
                  type: array
                  description: Optionally you can create lessons along with a group
                  items:
                    type: object
                    required:
                      - start
                      - end
                    properties:
                      start:
                        $ref: "#/components/schemas/Datetime"
                      end:
                        $ref: "#/components/schemas/Datetime"
  # # # # # # # #
  # FILE ROUTES #
  # # # # # # # #
...
components:
  schemas:
    Employee:
      type: object
      description: Employee information
      required:
        - id
        - created
        - name
        - email
        - roles
      properties:
        id:
          $ref: "#/components/schemas/ID"
        name:
          type: string
          example: John Doe
          maxLength: 48
          description: Full name of the employee
        email:
          type: string
          format: email
          example: john.doe@example.com
          description: Email address of the employee
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
          example:
            - teacher
            - administrator
          description: Roles assigned to the employee
        created:
          $ref: "#/components/schemas/Datetime"
    Location:
      type: object
      description: A location object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        created:
          $ref: "#/components/schemas/Datetime"
        name:
          type: string
          description: The name of the location
          example: Kecskeméti Kodály Zoltán Ének-Zenei Általános Iskola
        address:
          type: string
          description: The address of the location
          example: 6000 Kecskemét, Dózsa György út 22.
        contact_email:
          $ref: "#/components/schemas/Email"
        contact_phone:
          $ref: "#/components/schemas/PhoneNumber"
      required:
        - id
        - created
        - name
        - address
        - contact_email
        - contact_phone
...
    Group:
      type: object
      description: A group object
      required:
        - id
        - created
        - location
        - archived
        - teachers
      properties:
        id:
          $ref: "#/components/schemas/ID"
        created:
          $ref: "#/components/schemas/Datetime"
        location:
          $ref: "#/components/schemas/Location"
        archived:
          type: boolean
          description: Whether the group is archived or not
          example: false
        teachers:
          type: array
          items:
            anyOf:
              - $ref: "#/components/schemas/ID"
              - $ref: "#/components/schemas/Employee"
          description: The teachers assigned to the group
...
    Lesson:
      type: object
      description: A lesson object
      required:
        - id
        - created
        - name
        - group
        - location
        - teachers
        - start
        - end
      properties:
        id:
          $ref: "#/components/schemas/ID"
        created:
          $ref: "#/components/schemas/Datetime"
        name:
          type: string
          description: The name of the lesson
          example: Scratch lesson 1
        group:
          $ref: "#/components/schemas/Group"
        location:
          $ref: "#/components/schemas/Location"
        teachers:
          type: array
          items:
            anyOf:
              - $ref: "#/components/schemas/ID"
              - $ref: "#/components/schemas/Employee"
        start:
          $ref: "#/components/schemas/Datetime"
        end:
          $ref: "#/components/schemas/Datetime"
...
    Email:
      type: string
      format: email
      description: A string in email address format.
    PhoneNumber:
      type: string
      format: telephone
      example: "+36301131313"
      description: A telephone number starting with country code
    ID:
      type: string
      description: A surrealdb id in the format `table:id`
      example: sometable:1234
    Role:
      type: string
      description: A role is a string that can either be "administrator" or "teacher". More roles may be added in the future.
      enum:
        - administrator
        - teacher
      example: administrator
    Datetime:
      type: string
      description: ISO 8601 date-time string
      format: date-time
      example: 2023-10-15T14:30:00Z
  responses:
...
    BadRequest:
      description: Something is missing or incorrect in the client's request, so the server cannot process it.
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - code
            properties:
              code:
                enum:
                  - server_error
                  - bad_request
                  - unauthorized
                  - invalid_credentials
                  - invalid_request_body
                  - fields_required
                  - fields_invalid
                  - password_too_weak
                  - email_in_use
                  - invalid_invite
                type: string
                example: bad_request
              message:
                type: string
                example: Some error message...
    Unauthorized:
      description: Authentication is required to access the requested resource. The client must include the appropriate credentials.
      content:
        application/json:
          schema:
            type: object
            required:
              - message
              - code
            properties:
              code:
                enum: 
                  - unauthorized
                type: string
                example: unauthorized
              message:
                type: string
                example: You are not authorized to perform this action
  parameters: 
    ids:
      name: ids
      description: Ids in surrealdb format (table:id) separated by commas
      in: query
      style: form
      schema:
        $ref: "#/components/schemas/ID"
      explode: true
      required: true
      allowReserved: false
  securitySchemes:
    BearerJwt:
      type: http
      scheme: bearer
      description: This is the token used for authenticating users
      bearerFormat: Json Web Token (JWT)
...